requirements.txt
streamlit==1.29.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
Pillow==10.0.1
PyYAML==6.0.1
pydantic==2.5.0
reportlab==4.0.7
pytest==7.4.3
torch==2.1.1
onnxruntime==1.16.3
pandas==2.1.4
matplotlib==3.8.2